refactor: put the parsing stuff into 'render'
