+++ render/model.go
@@ -1,0 +1,4 @@
+package render
+
+var SeparatorOpen = "{"
+var SeparatorClose = "}"
