+++ sources/config.go
@@ -13,7 +13,7 @@ type Config struct {
 }
 
 func LoadConfig() error {
-	data, err := os.ReadFile(filepath.Join(dataPath, "config.json"))
+	data, err := os.ReadFile(filepath.Join(DataPath, "config.json"))
 	if err != nil && errors.Is(err, os.ErrNotExist) {
 		log.Printf("load config file err: %v, default settings are used.", err)
 		return err
@@ -26,8 +26,8 @@ func LoadConfig() error {
 	if len(c.Dicts) == 0 {
 		return nil
 	}
-	GlobalDict.mdxFile = c.Dicts[0]
-	GlobalDict.mdxCss = c.Dicts[0] + ".css"
+	GlobalDict.MdxFile = c.Dicts[0]
+	GlobalDict.MdxCss = c.Dicts[0] + ".css"
 	log.Printf("get global dicts: %v", GlobalDict)
 	return nil
 }
