+++ sources/model.go
@@ -9,7 +9,7 @@ import (
 
 var mu sync.Mutex // owns history
 var history map[string]string = make(map[string]string)
-var dataPath string
+var DataPath string
 var historyFile string
 
 func init() {
@@ -18,10 +18,10 @@ func init() {
 		log.Fatal(err)
 	}
 	configPath := filepath.Join(home, ".config")
-	dataPath = filepath.Join(configPath, "ondict")
-	historyFile = filepath.Join(dataPath, "history.json")
-	if dataPath == "" || historyFile == "" {
-		log.Fatalf("empty datapath/historyfile: %v||%v", dataPath, historyFile)
+	DataPath = filepath.Join(configPath, "ondict")
+	historyFile = filepath.Join(DataPath, "history.json")
+	if DataPath == "" || historyFile == "" {
+		log.Fatalf("empty datapath/historyfile: %v||%v", DataPath, historyFile)
 	}
 }
 
@@ -53,13 +53,13 @@ func (o output) GetDefinition() string {
 }
 
 func (d *MdxDict) Register() error {
-	d.mdxDict = loadDecodedMdx(filepath.Join(dataPath, "dicts", d.mdxFile))
-	if contents, err := os.ReadFile((filepath.Join(dataPath, "dicts", d.mdxCss))); err == nil {
-		d.mdxCss = string(contents)
+	d.MdxDict = loadDecodedMdx(filepath.Join(DataPath, "dicts", d.MdxFile))
+	if contents, err := os.ReadFile((filepath.Join(DataPath, "dicts", d.MdxCss))); err == nil {
+		d.MdxCss = string(contents)
 	} else {
-		d.mdxCss = ""
-		log.Printf("load dicts[%v] css err: %v", d.mdxFile, err)
+		d.MdxCss = ""
+		log.Printf("load dicts[%v] css err: %v", d.MdxFile, err)
 	}
-	d.searcher = New(d.mdxDict)
+	d.searcher = New(d.MdxDict)
 	return nil
 }
